--- client/src/main/java/com/sparta/rooibos/client/application/dto/req/UpdateHubIdRequest.java
@@ -1,3 +1,0 @@
-package com.sparta.rooibos.client.application.dto.req;
-
-public record UpdateHubIdRequest(String hubId) {}
