+++ client/src/main/java/com/sparta/rooibos/client/application/dto/request/CreateClientRequest.java
@@ -1,3 +1,3 @@
-package com.sparta.rooibos.client.application.dto.req;
+package com.sparta.rooibos.client.application.dto.request;
 
 public record CreateClientRequest(String name, String clientType, String managedHubId, String address) {}
