+++ client/src/main/java/com/sparta/rooibos/client/application/dto/request/UpdateClientRequest.java
@@ -1,4 +1,4 @@
-package com.sparta.rooibos.client.application.dto.req;
+package com.sparta.rooibos.client.application.dto.request;
 
 
 public record UpdateClientRequest(String name, String address) {}
