+++ client/src/main/java/com/sparta/rooibos/client/application/dto/request/UpdateHubIdRequest.java
@@ -1,0 +1,3 @@
+package com.sparta.rooibos.client.application.dto.request;
+
+public record UpdateHubIdRequest(String hubId) {}
