+++ client/src/main/java/com/sparta/rooibos/client/application/dto/response/GetClientResponse.java
@@ -1,4 +1,4 @@
-package com.sparta.rooibos.client.application.dto.res;
+package com.sparta.rooibos.client.application.dto.response;
 
 
 import java.time.LocalDateTime;
