+++ client/src/main/java/com/sparta/rooibos/client/application/dto/response/SearchClientListResponse.java
@@ -1,4 +1,4 @@
-package com.sparta.rooibos.client.application.dto.res;
+package com.sparta.rooibos.client.application.dto.response;
 
 import java.util.UUID;
